refactored lib functions - splitted into smaller chunks
