+++ src/lib/rectangleCoordinates.spec.js
@@ -3,8 +3,7 @@ import {
   getLeftCoordinates,
   getTopCoordinates,
   getRightCoordinates,
-  getZoomContainerDistance,
-  cropImage
+  getHorizontalContainerDistance
 } from './coordinates';
 
 describe('coordinates module', () => {
@@ -57,30 +56,6 @@ describe('coordinates module', () => {
         left: 10,
         top: 10
       });
-    });
-  });
-  describe('cropImage function', () => {
-    it('return margin coordinates which allow to crop image', () => {
-      const image = {
-        clientHeight: 300,
-        naturalHeight: 900,
-        naturalWidth: 1600
-      };
-      expect(cropImage(image, currentPositions)).toEqual({
-        marginTop: -30,
-        marginLeft: -30,
-        marginRight: -1525,
-        marginBottom: -840
-      });
-    });
-  });
-  describe('getZoomContainerDistance function', () => {
-    it('return distance between thumbnail and real image', () => {
-      const image = {
-        naturalHeight: 900,
-        clientHeight: 500
-      };
-      expect(getZoomContainerDistance(image, currentPositions)).toEqual(-47);
     });
   });
 });
